backend/pyproject.toml
[project]
name = "backlog-boss-backend"
version = "0.1.0"
description = "Backend for Backlog Boss"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "expiring-dict>=1.1.2",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "igdb-api-v4>=0.3.3",
    "pydantic-settings>=2.12.0",
    "pyodbc>=5.3.0",
    "python-steam-api>=2.2.1",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
    "testcontainers[mssql]>=4.14.2",
]

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
omit = ["*/__init__.py"]
